==> train/3_text_summarization_threshold_filtering_2.py <==
# -*- coding: utf-8 -*-
"""3_Text Summarization_Threshold Filtering_2

Automatically generated by Colaboratory.

Original file is located at
    https://colab.research.google.com/drive/1JKHEXBgm9X5vefGa-CQDSzigu-qrtrGN
"""

!pip install vaderSentiment

from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
import numpy as np  
import pandas as pd

### Initialization Variable### 
input_path = "/content/drive/MyDrive/PLP Practice Module/df_CA_Summarization_cleaned.csv"
input_sep = ","
output_path = '/content/drive/MyDrive/PLP Practice Module/df_CA_Summarization_filtered.csv'
output_sep = ","

# Load dataset 
def load_data():
  return pd.read_csv(input_path,sep=input_sep)

def filltering_main():
  
  data =load_data()

  ###VADER SENTIMENT MODEL### 
  vader_obj = SentimentIntensityAnalyzer()
  f_VADER = lambda cleaned_summary: vader_obj.polarity_scores(cleaned_summary)['compound'] 
  data['sentiment_VADER'] = data['cleaned_summary'].apply(f_VADER) 
  #data[['sentiment_VADER','cleaned_summary']].sample(30)

  ###APPLY FILTERING### 
  data=data[data['sentiment_VADER'] > 0.5]

  ###OUTPUT to CSV### 
  data.to_csv(output_path, sep=output_sep)

# call filltering_main()
filltering_main()
